PropertyFileUtil improvements
- PropertyFileUtilFileKindEnum class file created
- Unused methods (message related files without locale argument) removed
