+++ ecuacion-lib-core/src/test/java/jp/ecuacion/lib/core/util/internal/PropertyFileUtilValueGetterTest.java
@@ -15,13 +15,13 @@
  */
 package jp.ecuacion.lib.core.util.internal;
 
-import static jp.ecuacion.lib.core.util.PropertyFileUtil.PropertyFileUtilFileKindEnum.APPLICATION;
-import static jp.ecuacion.lib.core.util.PropertyFileUtil.PropertyFileUtilFileKindEnum.MESSAGES;
+import static jp.ecuacion.lib.core.util.enums.PropertyFileUtilFileKindEnum.APPLICATION;
+import static jp.ecuacion.lib.core.util.enums.PropertyFileUtilFileKindEnum.MESSAGES;
 
 import java.util.Locale;
 import jp.ecuacion.lib.core.TestTools;
 import jp.ecuacion.lib.core.util.ObjectsUtil.RequireNonNullException;
-import jp.ecuacion.lib.core.util.PropertyFileUtil.PropertyFileUtilFileKindEnum;
+import jp.ecuacion.lib.core.util.enums.PropertyFileUtilFileKindEnum;
 import jp.ecuacion.lib.core.util.internal.PropertyFileUtilValueGetter.KeyDupliccatedException;
 import jp.ecuacion.lib.core.util.internal.PropertyFileUtilValueGetter.NoKeyInPropertiesFileException;
 import org.junit.jupiter.api.Assertions;
